requirements.txt
et-xmlfile==1.1.0
lxml==5.3.0
numpy==2.1.2
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
pdf2image==1.17.0
pillow==10.4.0
pypdf==5.0.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2024.2
six==1.16.0
typing_extensions==4.12.2
tzdata==2024.2

